feat(attachment-uploader): move auth related stuff (authLoading, Context etc) from step to form
